clippy fixes

+++ quadratic-core/src/sheet_offsets/mod.rs
@@ -218,11 +218,11 @@ impl SheetOffsets {
     }
 
     pub fn total_column_width(&self, start: i64, end: i64) -> f64 {
-        self.column_widths.size(start as i64, end as i64)
+        self.column_widths.size(start, end)
     }
 
     pub fn total_row_height(&self, start: i64, end: i64) -> f64 {
-        self.row_heights.size(start as i64, end as i64)
+        self.row_heights.size(start, end)
     }
 }
